+++ crates/matrix-sdk-ui/src/timeline/tests/mod.rs
@@ -54,6 +54,7 @@ use super::{
     traits::RoomDataProvider,
     EventTimelineItem, Profile, TimelineInner, TimelineItem,
 };
+use crate::unable_to_decrypt_hook::UtdHookManager;
 
 mod basic;
 mod echo;
@@ -62,7 +63,6 @@ mod edit;
 mod encryption;
 mod event_filter;
 mod invalid;
-mod pagination;
 mod polls;
 mod reaction_group;
 mod reactions;
@@ -81,7 +81,17 @@ impl TestTimeline {
     }
 
     fn with_room_data_provider(room_data_provider: TestRoomDataProvider) -> Self {
-        Self { inner: TimelineInner::new(room_data_provider), event_builder: EventBuilder::new() }
+        Self {
+            inner: TimelineInner::new(room_data_provider, None),
+            event_builder: EventBuilder::new(),
+        }
+    }
+
+    fn with_unable_to_decrypt_hook(hook: Arc<UtdHookManager>) -> Self {
+        Self {
+            inner: TimelineInner::new(TestRoomDataProvider::default(), Some(hook)),
+            event_builder: EventBuilder::new(),
+        }
     }
 
     fn with_settings(mut self, settings: TimelineInnerSettings) -> Self {
@@ -233,10 +243,7 @@ impl TestTimeline {
 
     async fn handle_back_paginated_custom_event(&self, event: Raw<AnyTimelineEvent>) {
         let timeline_event = TimelineEvent::new(event.cast());
-        self.inner
-            .handle_back_paginated_events(vec![timeline_event], Default::default())
-            .await
-            .unwrap();
+        self.inner.handle_back_paginated_events(vec![timeline_event]).await.unwrap();
     }
 
     async fn handle_read_receipts(