+++ crates/matrix-sdk-ui/tests/integration/timeline/sliding_sync.rs
@@ -258,14 +258,14 @@ async fn timeline_test_helper(
     // TODO: when the event cache handles its own cache, we can remove this.
     client
         .event_cache()
-        .add_initial_events(room_id, sliding_sync_room.timeline_queue().iter().cloned().collect())
+        .add_initial_events(
+            room_id,
+            sliding_sync_room.timeline_queue().iter().cloned().collect(),
+            sliding_sync_room.prev_batch(),
+        )
         .await?;
 
-    let timeline = Timeline::builder(&sdk_room)
-        .with_pagination_token(sliding_sync_room.prev_batch())
-        .track_read_marker_and_receipts()
-        .build()
-        .await?;
+    let timeline = Timeline::builder(&sdk_room).track_read_marker_and_receipts().build().await?;
 
     Ok(timeline.subscribe().await)
 }